+++ crates/hook/src/lib.rs
@@ -26,7 +26,7 @@ pub trait IngestHook: Send + Sync + 'static {
     fn on_batch_decoded(&self, _run_id: RunId, _batch: &MetricBatch) {}
 
     /// Called for each bucket that closed during aggregation.
-    fn on_buckets_closed(&self, _run_id: RunId, _key: &Metric, _tier: u64, _bucket: &Bucket) {}
+    fn on_buckets_closed(&self, _run_id: RunId, _key: &Metric, _tier: usize, _bucket: &Bucket) {}
 
     /// Called when a run transitions between states.
     fn on_run_status_change(&self, _run_id: RunId, _old: &RunStatus, _new: &RunStatus) {}
